@@ -2,26 +2,47 @@
 
 #include "core/Common.hh"
 
+#include <atomic>
 #include <vector>
 
 namespace sp {
-    class AssetManager;
-
     class Asset : public NonCopyable {
     public:
-        Asset(const string &path) : path(path) {}
+        Asset(const std::string &path = "") : path(path) {}
+
+        bool Valid() const {
+            return valid.test();
+        }
+
+        void WaitUntilValid() const {
+            while (!valid.test()) {
+                valid.wait(false);
+            }
+        }
 
         std::string String() const {
+            Assert(valid.test(), "Accessing buffer string on invalid asset");
             return std::string((char *)buffer.data(), buffer.size());
         }
 
-        const char *CharBuffer() const {
-            return (char *)buffer.data();
+        const uint8_t *Buffer() const {
+            Assert(valid.test(), "Accessing buffer data on invalid asset");
+            return buffer.data();
+        }
+
+        const size_t BufferSize() const {
+            Assert(valid.test(), "Accessing buffer size on invalid asset");
+            return buffer.size();
         }
 
         Hash128 Hash() const;
 
-        const string path;
+        const std::string path;
+
+    private:
+        std::atomic_flag valid;
         std::vector<uint8_t> buffer;
+
+        friend class AssetManager;
     };
 } // namespace sp
