@@ -3,12 +3,16 @@
 #include "core/Common.hh"
 
 #include <array>
+#include <atomic>
 #include <glm/glm.hpp>
-#include <tinygltf/tiny_gltf.h>
+#include <map>
+
+namespace tinygltf {
+    class FsCallbacks;
+    class Model;
+} // namespace tinygltf
 
 namespace sp {
-
-    // Forward declarations
     class Asset;
 
     enum TextureType {
@@ -20,16 +24,21 @@ namespace sp {
     };
 
     class Model : public NonCopyable {
-
     public:
         Model(const string &name) : name(name){};
-        Model(const string &name, shared_ptr<const Asset> asset, shared_ptr<tinygltf::Model> model)
-            : Model(name, model) {
-            this->asset = asset;
-        };
-        Model(const string &name, shared_ptr<tinygltf::Model> model);
+        Model(const string &name, std::shared_ptr<const Asset> asset) : name(name), asset(asset){};
 
         virtual ~Model();
+
+        bool Valid() const {
+            return valid.test();
+        }
+
+        void WaitUntilValid() const {
+            while (!valid.test()) {
+                valid.wait(false);
+            }
+        }
 
         struct Attribute {
             size_t vertexByteOffset;
@@ -43,13 +52,13 @@ namespace sp {
         };
 
         enum class DrawMode {
-            Points = TINYGLTF_MODE_POINTS,
-            Line = TINYGLTF_MODE_LINE,
-            LineLoop = TINYGLTF_MODE_LINE_LOOP,
-            LineStrip = TINYGLTF_MODE_LINE_STRIP,
-            Triangles = TINYGLTF_MODE_TRIANGLES,
-            TriangleStrip = TINYGLTF_MODE_TRIANGLE_STRIP,
-            TriangleFan = TINYGLTF_MODE_TRIANGLE_FAN,
+            Points = 0,
+            Line,
+            LineLoop,
+            LineStrip,
+            Triangles,
+            TriangleStrip,
+            TriangleFan,
         };
 
         struct Primitive {
@@ -60,32 +69,47 @@ namespace sp {
             std::array<Attribute, 5> attributes;
         };
 
-        const string name;
-        vector<Primitive> primitives;
+        const std::string name;
 
         bool HasBuffer(size_t index) const;
         const std::vector<unsigned char> &GetBuffer(size_t index) const;
         Hash128 HashBuffer(size_t index) const;
 
-        int FindNodeByName(string name) const;
-        string GetNodeName(int node) const;
+        int FindNodeByName(std::string name) const;
+        std::string GetNodeName(int node) const;
         glm::mat4 GetInvBindPoseForNode(int nodeIndex) const;
         std::vector<int> GetJointNodes() const;
 
-        std::shared_ptr<const tinygltf::Model> GetModel() const {
+        const std::shared_ptr<const tinygltf::Model> &GetGltfModel() const {
+            Assert(valid.test(), "Accessing gltf on invalid model");
             return model;
         }
 
-        vector<glm::mat4> bones;
+        const std::vector<Primitive> &Primitives() const {
+            Assert(valid.test(), "Accessing primitives on invalid model");
+            return primitives;
+        }
+
+        const std::vector<glm::mat4> &Bones() const {
+            Assert(valid.test(), "Accessing bones on invalid model");
+            return bones;
+        }
 
     private:
+        void PopulateFromAsset(std::shared_ptr<const Asset> asset, const tinygltf::FsCallbacks *fsCallbacks = nullptr);
         void AddNode(int nodeIndex, glm::mat4 parentMatrix);
 
-        std::shared_ptr<tinygltf::Model> model;
+        std::atomic_flag valid;
         std::shared_ptr<const Asset> asset;
+
+        std::shared_ptr<const tinygltf::Model> model;
+        std::vector<Primitive> primitives;
+        std::vector<glm::mat4> bones;
 
         // TODO: support more than one "skin" in a GLTF
         std::map<int, glm::mat4> inverseBindMatrixForJoint;
         int rootBone;
+
+        friend class AssetManager;
     };
 } // namespace sp
