@@ -1,8 +1,11 @@
 #pragma once
 
-#include "assets/Model.hh"
-
 #include <ecs/Components.hh>
+#include <glm/glm.hpp>
+
+namespace sp {
+    class Model;
+}
 
 namespace ecs {
     struct Renderable {
