@@ -182,7 +182,7 @@ namespace sp {
                                  ImGuiWindowFlags_AlwaysAutoResize;
 
         static shared_ptr<GpuTexture> logoTex = graphics.GetContext()->LoadTexture(
-            GAssets.LoadImageByPath("logos/sp-menu.png"));
+            GAssets.LoadImage("logos/sp-menu.png"));
         static ImVec2 logoSize(logoTex->GetWidth() * 0.5, logoTex->GetHeight() * 0.5);
 
         if (selectedScreen == MenuScreen::Splash) {
@@ -345,7 +345,7 @@ namespace sp {
             ImGui::Begin("MenuCredits", nullptr, flags);
 
             static shared_ptr<GpuTexture> frLogoTex = graphics.GetContext()->LoadTexture(
-                GAssets.LoadImageByPath("logos/credits-frustra.png"));
+                GAssets.LoadImage("logos/credits-frustra.png"));
             static ImVec2 frLogoSize(frLogoTex->GetWidth() * 0.5, frLogoTex->GetHeight() * 0.5);
 
             ImGui::BeginChild("CreditScroller", ImVec2(600, 600), false, ImGuiWindowFlags_NoScrollbar);
